+++ src/Ncodi/CodeAnalysis/Symbols/TypeSymbol.cs
@@ -3,7 +3,7 @@
     public sealed class TypeSymbol : Symbol
     {
         public static readonly TypeSymbol Int = new TypeSymbol("int");
-        public static readonly TypeSymbol Decimal = new TypeSymbol("decimal");
+        public static readonly TypeSymbol Double = new TypeSymbol("double");
         public static readonly TypeSymbol Bool = new TypeSymbol("bool");
         public static readonly TypeSymbol String = new TypeSymbol("string");
         public static readonly TypeSymbol Error = new TypeSymbol("?");
